added optimizeAnkiDeck function
